the post,put,get,delete tested

+++ isd-backend/src/main/java/com/example/isdbackend/repository/ProviderRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProviderRepository extends CrudRepository<Provider, Integer> {
+public interface ProviderRepository extends CrudRepository<Provider, Long> {
     List<Provider> findAll();
 }
