+++ isd-backend/src/main/java/com/example/isdbackend/service/UserService.java
@@ -42,7 +42,7 @@ public class UserService extends AbstractServiceCrud {
         return userRepository.findById(id).orElseThrow().getOrders().stream().filter(order -> order.isOrdered()).collect(Collectors.toList());
     }
 
-    public Iterable<Menu> getProviderMenus(Integer providerId){ return providerRepository.findById(providerId).orElseThrow().getMenus(); }
+    public Iterable<Menu> getProviderMenus(Long providerId){ return providerRepository.findById(providerId).orElseThrow().getMenus(); }
 
     public Iterable<Menu> filter(Long providerId,Long userId){
         Set<Order> orders = userRepository.findById(userId).orElseThrow().getOrders();
